More Warning suppression

+++ c/parallel/src/radix_sort.cu
@@ -230,7 +230,8 @@ radix_sort_runtime_tuning_policy get_policy(int /*cc*/, uint64_t key_size)
     256, 8, std::max(1, 1 * 4 / std::max(static_cast<int>(key_size), 4)), onesweep_radix_bits};
   constexpr agent_radix_sort_exclusive_sum_policy exclusive_sum_policy{256, onesweep_radix_bits};
 
-  const auto [onesweep_items_per_thread, onesweep_block_threads] = reg_bound_scaling(256, 21, key_size);
+  const auto [onesweep_items_per_thread, onesweep_block_threads] =
+    reg_bound_scaling(256, 21, static_cast<int>(key_size));
   // const auto [scan_items_per_thread, scan_block_threads]         = mem_bound_scaling(512, 23, key_size);
   const int scan_items_per_thread = 5;
   const int scan_block_threads    = 512;
@@ -238,9 +239,10 @@ radix_sort_runtime_tuning_policy get_policy(int /*cc*/, uint64_t key_size)
   const int downsweep_items_per_thread = 5;
   const int downsweep_block_threads    = 160;
   // const auto [alt_downsweep_items_per_thread, alt_downsweep_block_threads] = mem_bound_scaling(256, 16, key_size);
-  const int alt_downsweep_items_per_thread                             = 5;
-  const int alt_downsweep_block_threads                                = 256;
-  const auto [single_tile_items_per_thread, single_tile_block_threads] = mem_bound_scaling(256, 19, key_size);
+  const int alt_downsweep_items_per_thread = 5;
+  const int alt_downsweep_block_threads    = 256;
+  const auto [single_tile_items_per_thread, single_tile_block_threads] =
+    mem_bound_scaling(256, 19, static_cast<int>(key_size));
 
   constexpr bool is_onesweep = false;
